+++ CHANGELOG.md
@@ -95,6 +95,7 @@
 - Make end date in events optional @robgietema
 - Replace momentjs with dayjs @robgietema
 - Add basic crud operations for content rules @robgietema
+- Add content rules action and conditions registry @robgietema
 
 ### Bugfix